toradex: use the correct DRI device with KMS
The correct DRI device to be used with Qt varies with each
Toradex board. Fill the kms.conf with correct device based
on the selected target machine. The KMS is used with mx8,
but also when ever mainline bsp is used instead of NXP.

Pick-to: 6.2

+++ meta-boot2qt-distro/dynamic-layers/freescale-layer/recipes-qt/boot2qt-addons/default-qt-envs.bbappend
@@ -29,7 +29,10 @@
 
 FILESEXTRAPATHS:prepend := "${THISDIR}/${PN}:"
 
-SRC_URI:append:mx8 = " file://kms.conf"
+SRC_URI += "file://kms.conf.in"
+
+# Default DRI device to use with KMS
+DRI_DEVICE ?= "card0"
 
 do_configure:append() {
     echo "FB_MULTI_BUFFER=2" >> ${WORKDIR}/defaults
@@ -40,10 +43,19 @@ do_configure:append:mx8() {
     echo "QT_QPA_EGLFS_FORCE888=1" >> ${WORKDIR}/defaults
     echo "QT_QPA_EGLFS_KMS_ATOMIC=1" >> ${WORKDIR}/defaults
     echo "QT_QPA_EGLFS_KMS_CONFIG=/etc/kms.conf" >> ${WORKDIR}/defaults
+    sed -e 's/@DEVICE@/${DRI_DEVICE}/' ${WORKDIR}/kms.conf.in > ${WORKDIR}/kms.conf
 }
 
-do_install:append:mx8() {
-    install -m 0644 ${WORKDIR}/kms.conf ${D}${sysconfdir}/
+do_configure:append:use-mainline-bsp() {
+    echo "QT_QPA_EGLFS_KMS_ATOMIC=1" >> ${WORKDIR}/defaults
+    echo "QT_QPA_EGLFS_KMS_CONFIG=/etc/kms.conf" >> ${WORKDIR}/defaults
+    sed -e 's/@DEVICE@/${DRI_DEVICE}/' ${WORKDIR}/kms.conf.in > ${WORKDIR}/kms.conf
+}
+
+do_install:append() {
+    if [ -e ${WORKDIR}/kms.conf ]; then
+        install -m 0644 ${WORKDIR}/kms.conf ${D}${sysconfdir}/
+    fi
 }
 
 do_configure:append:mx8mm() {
